Don't extend Buffer prototype by default.
Make extension of Buffer.prototype opt-in.  The rationale for that is
that some of the buffertools methods conflict with the Buffer methods
of the same name.  Users now have to call `buffertools.extend()` if
they want the old behavior.

Bump the version number to 2.0.0 because this is a breaking change to
the API.